--- src/Symfony/Component/HttpKernel/Security/Authentication/AuthenticationSuccessHandler.php
@@ -1,1 +1,0 @@
-<?php
